+++ core/src/test/java/org/lflang/tests/compiler/RoundTripTests.java
@@ -14,6 +14,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.lflang.Target;
 import org.lflang.ast.FormattingUtils;
 import org.lflang.ast.IsEqual;
+import org.lflang.ast.ParsingUtils;
 import org.lflang.lf.Model;
 import org.lflang.tests.LFInjectorProvider;
 import org.lflang.tests.LFTest;
@@ -41,14 +42,15 @@ public class RoundTripTests {
   }
 
   private void run(Path file) throws Exception {
-    Model originalModel = LfParsingUtil.parse(file);
+    Model originalModel = ParsingUtils.parse(file);
     System.out.println(file);
     assertThat(originalModel.eResource().getErrors(), equalTo(emptyList()));
     // TODO: Check that the output is a fixed point
     final int smallLineLength = 20;
     final String squishedTestCase = FormattingUtils.render(originalModel, smallLineLength);
     final Model resultingModel =
-        LfParsingUtil.parseSourceAsIfInDirectory(file.getParent(), squishedTestCase);
+        ParsingUtils.parseSourceAsIfInDirectory(file.getParent(), squishedTestCase);
+    LfParsingUtil.checkValid("file in " + file.getParent(), resultingModel);
 
     assertThat(resultingModel.eResource().getErrors(), equalTo(emptyList()));
     Assertions.assertTrue(
